remove duplicate code, some renaming, first deploy
