+++ geometry/geometry_utils.php
@@ -1,32 +1,45 @@
 <?php
 
-function get_centroid($line)
-{
-  $EARTH_RADIUS = 6371000; //m
+require_once "Point.class.php";
+require_once "LineString.class.php";
 
-  $length = $line->greatCircleLength($EARTH_RADIUS);
+function get_length($line) {
+  $points = $line->getPoints();
+  $distance = 0;
+  for($i = 0; $i+1 < count($points); $i++)
+  {
+    $pos1 = $points[$i];
+    $pos2 = $points[$i+1];
+	$distance += get_distance($pos1, $pos2);
+  }
+  return $distance;
+}
+
+function get_center($line) {
+  $length = get_length($line);
 
   $halfDistance = $length / 2;
 
-  if($halfDistance == 0) return $line->pointN(1);
+  $points = $line->getPoints();
+  if($halfDistance == 0) return $points[0];
 
   $distance = 0;
-	for($i = 1; $i < $line->numPoints(); $i++)
+  for($i = 0; $i+1 < count($points); $i++)
   {
-    $pos1 = $line->pointN($i);
-		$pos2 = $line->pointN($i+1);
-		$segmentDistance = get_distance($pos1, $pos2, $EARTH_RADIUS);
-		$distance += $segmentDistance;
+    $pos1 = $points[$i];
+    $pos2 = $points[$i+1];
+    $segmentDistance = get_distance($pos1, $pos2);
+    $distance += $segmentDistance;
 
     if($distance > $halfDistance)
     {
       $ratio = ($distance - $halfDistance) / $segmentDistance;
-			$lat = $pos2->y() - $ratio * ($pos2->y() - $pos1->y());
-			$lon = $pos2->x() - $ratio * ($pos2->x() - $pos1->x());
-			return new Point($lon, $lat);
+      $lat = $pos2->y() - $ratio * ($pos2->y() - $pos1->y());
+      $lon = $pos2->x() - $ratio * ($pos2->x() - $pos1->x());
+      return new Point($lon, $lat);
     }
   }
-	return null;
+  return null;
 }
 
 function get_data_string($wkt) {
@@ -63,25 +76,22 @@ function parse_point($data_string) {
 /**
 * @return distance between two points in meters
 */
-function get_distance($pos1, $pos2, $EARTH_RADIUS)
-{
-	return $EARTH_RADIUS * distance(
-			deg2rad($pos1->y()),
-			deg2rad($pos1->x()),
-			deg2rad($pos2->y()),
-			deg2rad($pos2->x()
-			));
+function get_distance($pos1, $pos2) {
+  return 6371000 * distance(
+    deg2rad($pos1->y()),
+    deg2rad($pos1->x()),
+    deg2rad($pos2->y()),
+    deg2rad($pos2->x()
+    ));
 }
 
-
 // https://en.wikipedia.org/wiki/Great-circle_navigation#cite_note-2
-function distance($φ1, $λ1, $φ2, $λ2)
-{
-	$Δλ = $λ2 - $λ1;
+function distance($φ1, $λ1, $φ2, $λ2) {
+  $Δλ = $λ2 - $λ1;
 
-	$y = sqrt(pow(cos($φ2)*sin($Δλ), 2) + pow(cos($φ1)*sin($φ2) - sin($φ1)*cos($φ2)*cos($Δλ), 2));
-	$x = sin($φ1)*sin($φ2) + cos($φ1)*cos($φ2)*cos($Δλ);
-	return atan2($y, $x);
+  $y = sqrt(pow(cos($φ2)*sin($Δλ), 2) + pow(cos($φ1)*sin($φ2) - sin($φ1)*cos($φ2)*cos($Δλ), 2));
+  $x = sin($φ1)*sin($φ2) + cos($φ1)*cos($φ2)*cos($Δλ);
+  return atan2($y, $x);
 }
